LRAC-11218 Return double value of sessions per visitor

+++ modules/dxp/apps/osb/osb-asah/osb-asah-common/src/main/java/com/liferay/osb/asah/common/model/VisitorBehaviorMetric.java
@@ -58,7 +58,7 @@ public abstract class VisitorBehaviorMetric {
 
 			BigDecimal sessionsPerVisitor = _sessions.divide(_visitors);
 
-			return sessionsPerVisitor.longValue();
+			return sessionsPerVisitor.doubleValue();
 		}
 
 		return 0;
